add errors handle

+++ src/entities/post.js
@@ -1,3 +1,4 @@
+const ErrorService = require("../middleware/error/errorService");
 const Post = require("../models/post");
 
 async function getAllPosts() {
@@ -7,6 +8,9 @@ async function getAllPosts() {
 
 async function getPostById(id) {
   const post = await Post.query().findById(id);
+  if (!post) {
+    throw ErrorService.errorThrow(404);
+  }
   return post;
 }
 
@@ -15,14 +19,26 @@ async function createPost(post) {
 }
 
 async function updatePost(id, data) {
+  const post = await Post.query().findById(id);
+  if (!post) {
+    throw ErrorService.errorThrow(404);
+  }
   return Post.query().findById(id).patch(data);
 }
 
 async function removePostById(id) {
+  const post = await Post.query().findById(id);
+  if (!post) {
+    throw ErrorService.errorThrow(404);
+  }
   return Post.query().findById(id).delete();
 }
 
 async function getPostsByCategory(categoryId) {
+  const post = await Post.query().where({ categoryId }).select().first();
+  if (!post) {
+    throw ErrorService.errorThrow(404);
+  }
   return Post.query()
     .where({
       categoryId,
@@ -31,6 +47,10 @@ async function getPostsByCategory(categoryId) {
 }
 
 async function getPostsByAuthor(author) {
+  const post = await Post.query().where({ author }).select().first();
+  if (!post) {
+    throw ErrorService.errorThrow(404);
+  }
   return Post.query()
     .where({
       author,
